Add files via upload
BookBinders: Predicting Response with Logistic Regression
As a direct marketer of specialty books, the BookBinders Book Club has achieved steady
growth in their customer base. Yet while sales have grown steadily, profits began falling when the database got larger and when the company diversified its book selection and increased the
number of offers sent to customers. The falling profits have led Dave Lawton, BookBinders'
marketing director, to experiment with different database marketing approaches in order improve BookBinders' mailing yields and profits.
Dave began a series of live market tests, each involving a random sample of customers from
the database. An offer for the current book selection is sent to the sample and then the sample
customers' responses, either purchase or no purchase, are recorded and used to calibrate a
response model for the current offering. The response model's results are then used to "score" the remaining customers in the database and select customers from the full customer database for the 'rollout' mailing campaign.
Logistic Regression offers a powerful method for modeling response. Logistic regression is similar to linear regression - the key difference is that the dependent variable is binary (for
example, purchase or no purchase) rather than continuous. For each customer, logistic regression predicts a probability, between 0 and 1, of purchase or response, which can be used
for targeting and prediction decisions. Like linear regression, it can accommodate both continuous and categorical predictors, including interaction terms. Its use in database
marketing has grown as software becomes more readily available and as familiarity with the approach grows.
The company currently has 550,000 customers who are being mailed catalogs. Dave has just received a dataset containing the responses of a random sample of 50,000 customers to a
new offering from BookBinders titled "The Art History of Florence." Dave is eager to assess the
potential value of logistic regression as a method for predicting customer response and has asked you to complete the following analyses.